+++ packages/core/src/components/deprecated-text/LargeText.tsx
@@ -1,0 +1,9 @@
+import * as React from "react";
+import { Text, TextProps } from "../text/Text";
+
+/**
+ * @deprecated Please use `Text` instead.
+ */
+export const LargeText: React.FC<Omit<TextProps, "variant">> = (props) => {
+  return <Text size={"large"} {...props} />;
+};