+++ packages/theme/src/index.ts
@@ -1,1 +1,6 @@
-export const nothing = undefined;
+import "./styles/default-colors.css";
+import "./styles/default-theme.css";
+
+export * from "./util/CssColor";
+export * from "./types/ThemeColor";
+export * from "./types/LhdsColor";